refactor: keep two files under the 1000-line limit they just crossed
`storage.rs` (983 -> 1057) and `release_workflow_test.rs` (997 -> 1013) were
both a few lines under the repository's limit, and this branch's additions
pushed them over. The gate catches it; splitting is the fix.

`storage_budget.rs` takes the per-record budget arithmetic and record
reconciliation -- pure functions over a TokenRecord that decide whether a
request fits its bounds and how two copies reconcile, with no knowledge of
how records are stored.

`release_docker_test.rs` takes the three tests about what the release
workflow builds into the runtime image, plus the helpers only they use.

No behaviour changes. 52 suites pass on the 1.98.0 toolchain CI pins.
